Fix for compile/uncompiled options
Sometimes a report would be submitted as compiled even if the jasper file didn't exist, so I tweaked the logic to go ahead and compile it if the jasper file didn't exist.

Also, added a fonts.jar and tweaked the CL to modify the classpath to include the jar.  That way, in the future, a user can add fonts and create a jar file called fonts.jar and replace what is there to support their fonts